Update task for restructure
