+++ properties/task/base.py
@@ -4,8 +4,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from ..base import HasProperties
-from ..basic import String, Bool
+from ..hasproperties import HasProperties
+from ..primitive import Bool, String
 
 
 class BaseResult(HasProperties):
